+++ src/rule.js
@@ -15,7 +15,6 @@ class Rule extends EventEmitter {
    * @param {string} options.event.params - parameters to pass to the event listener
    * @param {Object} options.conditions - conditions to evaluate when processing this rule
    * @param {any} options.name - identifier for a particular rule, particularly valuable in RuleResult output
-   * @param {any} options.id - identifier for a particular rule, particularly valuable in RuleResult output
    * @return {Rule} instance
    */
   constructor (options) {
@@ -36,13 +35,6 @@ class Rule extends EventEmitter {
       this.setName(options.name)
     }
 
-    if (!options || options.id === undefined) {
-      // auto generate an id if none was provided
-      this.setId('_' + Math.random().toString(36).substr(2, 9))
-    } else {
-      this.setId(options.id)
-    }
-
     const priority = (options && options.priority) || 1
     this.setPriority(priority)
 
@@ -58,15 +50,6 @@ class Rule extends EventEmitter {
     priority = parseInt(priority, 10)
     if (priority <= 0) throw new Error('Priority must be greater than zero')
     this.priority = priority
-    return this
-  }
-
-  /**
-   * Sets the unique id of the rule
-   * @param {any} id - any truthy input
-   */
-  setId (id) {
-    this.id = id
     return this
   }